requirements.txt
httpx==0.27.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
playwright==1.53.0
pypdf==5.2.0
fastapi==0.116.1
uvicorn==0.35.0
